Small fixes (#3)
* Neaten up do {} syntax blocks

`return` inside a `do` block doesn't just return from the do, but makes
the entire containing function return. This probably wasn't what was wanted.

Also neaten up the indentation and add the trailing `;` to the final
statements.

* No trailing `;` on try {} catch {} blocks

As we're using Syntax::Keyword::Try there is no need to put a trailing
semicolon, as the module creates real statement blocks, not mere
expressions. This could be confused with Try::Tiny syntax if they remain.

* Remove redundant ->else call

This ->else block simply created another Future containing the same failure
as the first one. It can be removed with no change in meaning.

* Neaten ->transform call into async/await syntax

Use of a ->transform call to change the result of a Future can be neatened
up by the use of async/await syntax provided by Future::AsyncAwait instead.

* Perform both sending and reference address validation concurrently

No need to serialise these as they are performing independent RPC calls.
Might as well combine the two into a single Future->needs_all to make them
run concurrently.

* Further reduce copied code by using map{}

Since the two lines of code now differ only by the name of the hash key, we
might as well use map{} inside the Future->needs_all() call.